Game images
In your Arkanoid game (JavaScript version), each player starts with three lives. The objective of the game is to destroy all the bricks within the given level by controlling a paddle that bounces a ball. The game will continue until either all the bricks are destroyed or the player loses all their lives.

If the ball falls off the screen (due to the paddle missing it), the player will lose one life. The game ends when the player successfully clears the level by breaking all the bricks or loses all three lives.